Extract changelog entry component
